support logging

+++ toggl_jira_sync/main.py
@@ -1,7 +1,11 @@
+import argparse
 from collections import defaultdict
 from datetime import datetime, timedelta
 import logging
+import logging.config
 import operator
+
+from jira.utils import JIRAError
 
 from toggl_jira_sync.jira import (
     connect_to_jira,
@@ -13,14 +17,18 @@ from toggl_jira_sync.load_config import get_config
 from toggl_jira_sync.toggl import get_toggl_time_entries
 
 
-logger = logging.getLogger(__name__)
+logger = logging.getLogger('toggl_jira_sync')
+logging.config.fileConfig('logging.ini')
 
 
-def _get_jira_issue_ids_to_sync(config):
+def _get_jira_issue_ids_to_sync(config, lookback_days):
+    if not lookback_days:
+        lookback_days = 5
+
     toggl_entries = get_toggl_time_entries(
         config,
         {'since': (
-            datetime.now().date() - timedelta(days=5)
+            datetime.now().date() - timedelta(days=lookback_days)
         ).strftime('%Y-%m-%d')}
     )
 
@@ -48,18 +56,23 @@ def _get_comment(config, toggl_entries):
     ]) + ' minutes'
 
 
-def _sync_toggl_with_jira(config, jira_issue_id, jira_client):
+def _sync_toggl_with_jira(config, jira_issue_id, jira_client, dry_run=False):
     toggl_entries = get_toggl_time_entries(
         config,
         {'description': jira_issue_id.lower()}
     )
 
-    jira_worklogs = get_jira_worklogs(jira_client, jira_issue_id)
+    try:
+        jira_worklogs = get_jira_worklogs(jira_client, jira_issue_id)
+    except JIRAError:
+        logger.info('jira_issue_id {} does not exist in JIRA'.format(jira_issue_id))
+        return False
 
     toggl_total = sum([entry.seconds for entry in toggl_entries])
     jira_total = sum([entry.seconds for entry in jira_worklogs])
 
     if abs(toggl_total - jira_total) < 60:
+        logger.info('already_synced, {}'.format(jira_issue_id))
         return True
 
     jira_worklogs_toggl_user = [
@@ -69,11 +82,13 @@ def _sync_toggl_with_jira(config, jira_issue_id, jira_client):
     ]
 
     for worklog in jira_worklogs_toggl_user:
-        logger.info('delete_worklog', extra=dict(
-            jira_issue_id=worklog.jira_issue_id,
-            worklog_id=worklog.worklog_id,
+        logger.info('delete_worklog, {} - {}'.format(
+            worklog.jira_issue_id,
+            worklog.worklog_id,
         ))
-        delete_worklog(jira_client, worklog)
+        if not dry_run:
+            delete_worklog(jira_client, worklog)
+
         jira_total -= worklog.seconds
 
     seconds = toggl_total - jira_total
@@ -83,24 +98,41 @@ def _sync_toggl_with_jira(config, jira_issue_id, jira_client):
 
     comment = _get_comment(config, toggl_entries)
 
-    logger.info('add_worklog', extra=dict(
-        jira_issue_id=jira_issue_id,
-        seconds=seconds,
-    ))
-    add_worklog(
-        jira_client,
+    logger.info('add_worklog, {} - {} - {}'.format(
         jira_issue_id,
         seconds,
         comment,
-    )
+    ))
+    if not dry_run:
+        add_worklog(
+            jira_client,
+            jira_issue_id,
+            seconds,
+            comment,
+        )
 
     return True
 
 
-def main():
+def main(lookback_days, dry_run):
     config = get_config()
 
     jira_client = connect_to_jira(config)
 
-    for jira_issue_id in _get_jira_issue_ids_to_sync(config):
-        _sync_toggl_with_jira(config, jira_issue_id, jira_client)
+    for jira_issue_id in _get_jira_issue_ids_to_sync(config, lookback_days):
+        _sync_toggl_with_jira(config, jira_issue_id, jira_client, dry_run)
+
+
+def cli():
+    parser = argparse.ArgumentParser(
+        description='syncs Toggl with JIRA'
+    )
+    parser.add_argument('--dry-run', action='store_true')
+    parser.add_argument('--lookback-days')
+
+    cli_args = vars(parser.parse_args())
+
+    dry_run = cli_args['dry_run']
+    lookback_days = cli_args['lookback_days']
+
+    main(lookback_days, dry_run)
